Feat: 지역 도메인 생성
Member에 잘못된 설정이 존재하여 수정도 진행

Related to: #3

+++ src/main/java/com/ssafy/tott/global/domain/BaseEntity.java
@@ -18,12 +18,12 @@ import java.time.LocalDate;
 public abstract class BaseEntity {
     @CreatedBy
     @Column(nullable = false)
-    private long createdMemberId;
+    private int createdMemberId;
     @CreatedDate
     @Column(nullable = false)
     private LocalDate createdDate;
     @LastModifiedBy
-    private long updatedMemberId;
+    private int updatedMemberId;
     @LastModifiedDate
     private LocalDate updatedDate;
 }
